fix: fixed validations problems
