+++ apps/website/components/templates/protocol/data-models/show/components/data-model-tabs.tsx
@@ -55,6 +55,7 @@ export default function DataModelTabs({ dataModel, stats }: Props) {
             width="100%"
             height="800"
             frameBorder="0"
+            scrolling="no"
           ></iframe>
         </>
       ),
