+++ apps/website/locales/en/protocol.json
@@ -36,13 +36,10 @@
     "recipients": "Recipients",
     "issue-credential-title": "Issue Credential",
     "successfully-title": "Credential issued successfully",
-    "confirmation-dialog-title": "Confirmation",
-    "confirmation-dialog-text": "Are you sure you want to create this credential?",
     "error-on-create-credential": "An error occured, couldn't create the credential.",
+    "error-on-upload": "There was an error loading images.",
     "actions": {
       "issue-credential": "Issue Credential",
-      "confirmation-dialog-positive": "Create",
-      "confirmation-dialog-cancel": "Cancel",
       "check-credential": "Check credential"
     },
     "issue-credential": {
@@ -59,6 +56,8 @@
       "expire-date-label": "Add expire date",
       "group-claim-title": "Add details",
       "group-claim-description": "Add the details of the credential",
+      "group-issue-by-title": "Set the issuer",
+      "group-issue-by-description": "The issuer will be issued by",
       "issue-by": "Issue by"
     }
   }
